Allow Web Administrators to access the Privacy Settings admin page
This commit grants the `manage_privacy_options` capability to the Web Administrator role.

The admin menu is tweaked so that users with this role can still access the Privacy Settings page despite not having permission to `manage_options` (and therefore not having access to any other Settings pages). To do this, a new top-level 'Privacy' menu item is introduced for users without the `manage_options` capability.

A filter has also been applied to `map_meta_cap` which maps the capability `manage_privacy_options` to itself:

- Users with `manage_options` are automatically granted the `manage_privacy_options` capability. This is the default behaviour of WordPress.
- Users without `manage_options` capability may still be granted `manage_privacy_options` capability directly. This is implemented via the filter `Hooks::managePrivacyOptionsCap` in this commit.